# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loadout HTTP API",
    "version": "0.2.0",
    "description": "实际Go app/identity/运维路由。网页会话与ldt_网关令牌分离；业务写请求校验Origin；外部支付目前不可用。Body总大小app默认16KiB、技能发布50MiB（另验32文件/单文件256KiB）、identity4KiB；字符串长度限制以Go UTF-8字节校验为准，JSON Schema maxLength仅作客户端提示。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源服务根地址"
    }
  ],
...
  "security": [
    {
      "CookieSession": []
    }
  ],
  "paths": {
...
    "/api/v1/meta": {
      "get": {
        "operationId": "get_api_v1_meta",
        "summary": "外部能力配置状态",
        "tags": ["账号"],
        "description": "外部能力配置状态",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
...
    "/api/v1/account/email": {
      "get": {
        "operationId": "get_api_v1_account_email",
        "summary": "邮箱状态",
        "tags": ["账号"],
        "description": "邮箱状态",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStatus"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
    "/api/v1/account/email/request": {
      "post": {
        "operationId": "post_api_v1_account_email_request",
        "summary": "发送邮箱验证邮件",
        "tags": ["账号"],
        "description": "配置真实SMTP后发送；邮件不可用503 email_unavailable；限频429 rate_limited。",
        "responses": {
          "202": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": "sent"
                    }
                  },
                  "required": ["status"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 1,
                    "format": "email"
                  }
                },
                "required": ["email"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "post_api_v1_auth_email_verify",
        "summary": "确认邮箱验证链接",
        "tags": ["账号"],
        "description": "一次性链接，网页必须用户确认后提交，成功后清除地址栏token。无效/过期400 invalid_token。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": "verified"
                    }
                  },
                  "required": ["status"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "writeOnly": true
                  }
                },
                "required": ["token"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/github/start": {
      "get": {
        "operationId": "get_api_v1_auth_github_start",
        "summary": "开始GitHub授权",
        "tags": ["账号"],
        "description": "配置GitHub OAuth才启用；未配置503 github_unavailable。使用PKCE与浏览器state cookie，不将GitHub access token存入用户凭证。",
        "responses": {
          "302": {
            "description": "重定向至GitHub授权页或登录成功后的站点首页。",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "start设置10分钟state cookie；callback清state并签发AT/RT，期限由共享系统配置读取。"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/github/callback": {
      "get": {
        "operationId": "get_api_v1_auth_github_callback",
        "summary": "GitHub回调",
        "tags": ["账号"],
        "description": "配置GitHub OAuth才启用；未配置503 github_unavailable。使用PKCE与浏览器state cookie，不将GitHub access token存入用户凭证。",
        "responses": {
          "302": {
            "description": "重定向至GitHub授权页或登录成功后的站点首页。",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "start设置10分钟state cookie；callback清state并签发AT/RT，期限由共享系统配置读取。"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_session",
        "description": "HttpOnly、SameSite=Lax，默认900秒opaque AT；生产Secure。旧7天session在原期限内兼容。"
      },
      "GatewayToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ldt_ token",
        "description": "网页会话不能替代此令牌。"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_refresh"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1,
            "description": "稳定错误码；没有原始响应或内部凭证。"
          }
        },
        "required": ["error"]
      },
...
      "Meta": {
        "type": "object",
        "properties": {
          "github": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          },
          "payments": {
            "const": false
          }
        },
        "required": ["github", "email", "payments"]
      },
...
      "EmailStatus": {
        "type": "object",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "configured": {
            "type": "boolean"
          }
        },
        "required": ["email", "verified_at", "configured"]
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "业务错误；参见 error 稳定代码。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}